pyproject.toml
[tool.poetry]
name = "sirius-app"
version = "0.1.0"
description = "Сервис для регистрации и вопросов жителей ФТ Сириус"
authors = ["Your Name <email@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110"
uvicorn = "^0.29"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
python-jose = {extras = ["cryptography"], version = "^3.3"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
fastapi-mail = "^1.4"
pydantic = "^2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
